+++ changelog.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+dev
+-----------------
+- add CrossReference class
+- add cross_references attr to Enactment model
+
 0.2.0 (2020-08-30)
 ------------------
 - don't add ellipsis to selected_text for node with no text